+++ src/multiLine/gallery.ts
@@ -27,9 +27,7 @@ import {cached} from '../../mixin/cached';
 /* NOT FOR BROWSER END */
 
 /**
- * gallery tag
- *
- * gallery标签
+ * `<gallery>`
  * @classdesc `{childNodes: (GalleryImageToken|CommentLineToken|AstText)[]}`
  */
 export abstract class GalleryToken extends MultiLineToken {