+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [`AttributesToken.prototype.css`](https://github.com/bhsd-harry/wikiparser-node/wiki/AttributesToken-%28EN%29#css) now inserts the `style` attribute if not present
 - [`ParamLineToken.prototype.lint`](https://github.com/bhsd-harry/wikiparser-node/wiki/ParamLineToken-%28EN%29#lint) now reports unknown parameter names
 - [`LanguageService.prototype.provideCompletionItems`](https://github.com/bhsd-harry/wikiparser-node/wiki/LanguageService-%28EN%29#providecompletionitems) now suggests parameter names in `<dynamicpagelist>` and `<inputbox>`
+- Parse and report extension tags inside `<categorytree>`
 
 **Fixed**